src/requirements.txt
# Данный проект умеет собирать ссылки с сайта Википедии, визуализировать их в виде графов,
# искать самые короткие пути между узлами, а также создавать изображения этих графов
# в форматах .html и .png.

# Чтобы насладиться всеми возможностями этого проекта,
# настоятельно рекомендуется создать виртуальное окружение.

# Чтобы погрузиться в виртуальное окружение, выполните следующие шаги:
# 1. Создайте виртуальное окружение:
#     python -m venv myenv
# 2. Активируйте виртуальное окружение:
#     source myenv/bin/activate (для Unix/Mac)
#     myenv\Scripts\activate (для Windows)
# 3. Установите необходимые библиотеки:
#     pip install -r requirements.txt
# После установки всех необходимых библиотек, всё готово к работе!
# Для выхода из виртуального окружения выполните команду:
#     deactivate

beautifulsoup4==4.12.2
altair==5.1.2
bs4==0.0.1
networkx==2.6.3
nx-altair==0.1.0
pandas==2.1.2
requests==2.31.0
python-dotenv
matplotlib
scipy
